requirements.txt
pycryptodome>=3.20.0
pycryptodomex>=3.20.0